[FIX] Fixed ApiCamera due to new twincat structures & first test with both cameras. ApiADS is broken for now due to WIP TwinCat.

+++ Core/Entities/Packets/Models/Structs/RIDStruct.cs
@@ -6,7 +6,7 @@ namespace Core.Entities.Packets.Models.Structs;
 [StructLayout(LayoutKind.Sequential, Pack = 0, CharSet = CharSet.Ansi)]
 public struct RIDStruct
 {
-	public ushort StationID { get; set; }
+	public int StationID { get; set; }
 	public TimestampStruct TS { get; set; }
 
 	public string ToRID()
